Split LabelEditor.py into LabelEditorParts/
