+++ src/windows/mainWindow/elements/Sidebar/elements/LabelEditorParts/LabelGroup.py
@@ -1,0 +1,41 @@
+"""
+Author: Core447
+Year: 2023
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+any later version.
+
+This programm comes with ABSOLUTELY NO WARRANTY!
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <https://www.gnu.org/licenses/>.
+"""
+import gi
+
+from src.backend.DeckManagement.InputIdentifier import InputIdentifier
+from src.windows.mainWindow.elements.Sidebar.elements.LabelEditorParts.LabelExpanderRow import (
+    LabelExpanderRow,
+)
+
+gi.require_version("Gtk", "4.0")
+gi.require_version("Adw", "1")
+from gi.repository import Adw
+
+
+class LabelGroup(Adw.PreferencesGroup):
+    def __init__(self, sidebar, **kwargs):
+        super().__init__(**kwargs)
+        self.sidebar = sidebar
+
+        self.build()
+
+    def build(self):
+        self.expander = LabelExpanderRow(self)
+        self.add(self.expander)
+
+        return
+
+    def load_for_identifier(self, identifier: InputIdentifier, state: int):
+        self.expander.load_for_identifier(identifier, state)
