+++ src/windows/mainWindow/elements/Sidebar/elements/LabelEditorParts/LabelRow.py
@@ -1,0 +1,361 @@
+"""
+Author: Core447
+Year: 2023
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+any later version.
+
+This programm comes with ABSOLUTELY NO WARRANTY!
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <https://www.gnu.org/licenses/>.
+"""
+import threading
+
+import gi
+from loguru import logger as log
+
+import globals as gl
+from src.backend.DeckManagement.DeckController import KeyLabel
+from src.backend.DeckManagement.HelperMethods import (
+    color_values_to_gdk,
+    gdk_color_to_values,
+    get_pango_font_description,
+    get_values_from_pango_font_description,
+)
+from src.backend.DeckManagement.InputIdentifier import InputIdentifier
+from src.windows.mainWindow.elements.Sidebar.elements.LabelEditorParts.widgets import (
+    AlignmentButtons,
+    ColorChooserButton,
+    FontChooserButton,
+    SpinButton,
+    TextEntry,
+)
+
+gi.require_version("Gtk", "4.0")
+gi.require_version("Adw", "1")
+from gi.repository import Adw, Gtk
+
+
+class LabelRow(Adw.PreferencesRow):
+    def __init__(self, label_text, label_index: int, sidebar, key_name: str, **kwargs):
+        super().__init__(**kwargs)
+        self.label_text = label_text
+        self.sidebar = sidebar
+        self.active_identifier: InputIdentifier = None
+        self.state: int = 0
+        self.label_index = label_index
+        self.key_name = key_name
+        self.build()
+
+        self.lock = threading.Lock()
+
+        self.block = False
+
+        # Connect set signals
+        self.connect_signals()
+
+    def build(self):
+        self.main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, hexpand=True,
+                                margin_start=15, margin_end=15, margin_top=15, margin_bottom=15)
+        self.set_child(self.main_box)
+
+        self.label = Gtk.Label(label=self.label_text, xalign=0, margin_bottom=3, css_classes=["bold"])
+        self.main_box.append(self.label)
+
+        self.controlled_by_action_label = Gtk.Label(label=gl.lm.get("label-editor-warning-controlled-by-action"), css_classes=["bold", "red-color"], xalign=0,
+                                                    margin_bottom=3, visible=False)
+        self.main_box.append(self.controlled_by_action_label)
+
+        self.text_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, hexpand=True)
+        self.main_box.append(self.text_box)
+
+        self.text_entry = TextEntry()
+        self.text_box.append(self.text_entry)
+
+        self.color_chooser_button = ColorChooserButton()
+        self.text_box.append(self.color_chooser_button)
+
+        self.font_chooser_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, hexpand=True, margin_top=6)
+        self.main_box.append(self.font_chooser_box)
+
+        self.font_chooser_label = Gtk.Label(label=gl.lm.get("label-editor-font-chooser-label"), xalign=0, hexpand=True, margin_start=2)
+        self.font_chooser_box.append(self.font_chooser_label)
+
+        self.font_chooser_button = FontChooserButton()
+        self.font_chooser_box.append(self.font_chooser_button)
+
+        # Alignment box
+        self.alignment_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, hexpand=True, margin_top=6)
+        self.main_box.append(self.alignment_box)
+
+        self.alignment_label = Gtk.Label(label=gl.lm.get("label-editor-alignment-label", "Align:"), xalign=0, hexpand=True, margin_start=2)
+        self.alignment_box.append(self.alignment_label)
+
+        self.alignment_buttons = AlignmentButtons()
+        self.alignment_box.append(self.alignment_buttons)
+
+        self.outline_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, hexpand=True, margin_top=6)
+        self.main_box.append(self.outline_box)
+
+        self.outline_width_label = Gtk.Label(label=gl.lm.get("label-editor-outline-width-label"), xalign=0, hexpand=True, margin_start=2)
+        self.outline_width_label.set_hexpand(False)
+        self.outline_width_label.set_margin_end(5)
+        self.outline_box.append(self.outline_width_label)
+
+        self.outline_width = SpinButton(0, 10, 1)
+        self.outline_width.set_hexpand(False)
+        self.outline_box.append(self.outline_width)
+
+        self.outline_color_label = Gtk.Label(label=gl.lm.get("label-editor-outline-color-label"), xalign=0, hexpand=True, margin_start=2)
+        self.outline_color_label.set_hexpand(True)
+        self.outline_color_label.set_margin_end(5)
+        self.outline_color_label.set_halign(Gtk.Align.END)
+        self.outline_box.append(self.outline_color_label)
+
+        self.outline_color_chooser_button = ColorChooserButton()
+        self.outline_color_chooser_button.set_hexpand(False)
+        self.outline_box.append(self.outline_color_chooser_button)
+
+        ## Connect reset buttons
+        self.text_entry.revert_button.connect("clicked", self.on_reset_text)
+        self.color_chooser_button.revert_button.connect("clicked", self.on_reset_color)
+        self.font_chooser_button.revert_button.connect("clicked", self.on_reset_font)
+        self.outline_width.revert_button.connect("clicked", self.on_reset_outline_width)
+        self.outline_color_chooser_button.revert_button.connect("clicked", self.on_reset_outline_color)
+        self.alignment_buttons.revert_button.connect("clicked", self.on_reset_alignment)
+
+    def connect_signals(self):
+        self.text_entry.entry.connect("changed", self.on_change_text)
+        self.color_chooser_button.button.connect("color-set", self.on_change_color)
+        self.font_chooser_button.button.connect("font-set", self.on_change_font)
+        self.outline_width.button.connect("value-changed", self.on_change_outline_width)
+        self.outline_color_chooser_button.button.connect("color-set", self.on_change_outline_color)
+        self.alignment_buttons.left_button.connect("toggled", self.on_change_alignment)
+        self.alignment_buttons.center_button.connect("toggled", self.on_change_alignment)
+        self.alignment_buttons.right_button.connect("toggled", self.on_change_alignment)
+
+    def disconnect_signals(self):
+        try:
+            self.text_entry.entry.disconnect_by_func(self.on_change_text)
+        except Exception as e:
+            log.error(f"Failed to disconnect signals. Error: {e}")
+
+        try:
+            self.color_chooser_button.button.disconnect_by_func(self.on_change_color)
+        except Exception as e:
+            log.error(f"Failed to disconnect signals. Error: {e}")
+
+        try:
+            self.font_chooser_button.button.disconnect_by_func(self.on_change_font)
+        except Exception as e:
+            log.error(f"Failed to disconnect signals. Error: {e}")
+
+        try:
+            self.outline_width.button.disconnect_by_func(self.on_change_outline_width)
+        except Exception as e:
+            log.error(f"Failed to disconnect signals. Error: {e}")
+
+        try:
+            self.outline_color_chooser_button.button.disconnect_by_func(self.on_change_outline_color)
+        except Exception as e:
+            log.error(f"Failed to disconnect signals. Error: {e}")
+
+        try:
+            self.alignment_buttons.left_button.disconnect_by_func(self.on_change_alignment)
+            self.alignment_buttons.center_button.disconnect_by_func(self.on_change_alignment)
+            self.alignment_buttons.right_button.disconnect_by_func(self.on_change_alignment)
+        except Exception as e:
+            log.error(f"Failed to disconnect signals. Error: {e}")
+
+    def load_for_identifier(self, identifier: InputIdentifier, state: int):
+        if not isinstance(identifier, InputIdentifier):
+            raise ValueError
+        self.active_identifier = identifier
+        self.state = state
+
+        controller = gl.app.main_win.get_active_controller()
+        if controller is None:
+            return
+        page = controller.active_page
+
+        if page == None:
+            #TODO: Show error
+            return
+        
+        controller_input = controller.get_input(identifier)
+        use_page_label_properties = controller_input.get_active_state().label_manager.get_use_page_label_properties(position=self.key_name)
+
+        ## Set visibility of revert buttons
+        self.text_entry.revert_button.set_visible(use_page_label_properties.get("text", False))
+        self.color_chooser_button.revert_button.set_visible(use_page_label_properties.get("color", False))
+        self.outline_width.revert_button.set_visible(use_page_label_properties.get("outline_width", False))
+        self.outline_color_chooser_button.revert_button.set_visible(use_page_label_properties.get("outline_color", False))
+        self.alignment_buttons.revert_button.set_visible(use_page_label_properties.get("alignment", False))
+
+        font_combined = use_page_label_properties.get("font-family", False) and use_page_label_properties.get("font-size", False)
+        self.font_chooser_button.revert_button.set_visible(font_combined)
+
+        # Set properties
+        self.update_values()
+
+    def update_values(self, composed_label: KeyLabel = None):
+        self.lock.acquire()
+        self.disconnect_signals()
+        if composed_label is None:
+            controller = gl.app.main_win.get_active_controller()
+            controller_input = controller.get_input(self.active_identifier)
+            composed_label = controller_input.get_active_state().label_manager.get_composed_label(position=self.key_name)
+
+        if self.text_entry.entry.get_text() != composed_label.text:
+            pos = self.text_entry.entry.get_position()
+            
+            self.text_entry.entry.set_text(composed_label.text)
+
+            pos = min(pos, len(composed_label.text))
+            self.text_entry.entry.set_position(pos)
+
+        hide_details = composed_label.text.strip() == ""
+        self.font_chooser_box.set_visible(not hide_details)
+        self.outline_box.set_visible(not hide_details)
+        self.alignment_box.set_visible(not hide_details)
+
+        self.set_color(composed_label.color)
+        self.set_outline_width(composed_label.outline_width)
+        self.set_outline_color(composed_label.outline_color)
+        self.set_alignment(composed_label.alignment)
+
+        # self.font_chooser_button.button.set_font_desc(Pango.FontDescription.from_string(f"{composed_label.font_name} {composed_label.style} {composed_label.font_size}px"))
+        desc = get_pango_font_description(
+            font_family=composed_label.font_name,
+            font_size=composed_label.font_size,
+            font_style=composed_label.style,
+            font_weight=composed_label.font_weight
+        )
+        self.font_chooser_button.button.set_font_desc(desc)
+
+        self.connect_signals()
+
+        self.lock.release()
+
+    def set_color(self, color_values: list):
+        color = color_values_to_gdk(color_values)
+        self.color_chooser_button.button.set_rgba(color)
+
+    def set_outline_width(self, outline_width: int):
+        self.outline_width.button.set_value(outline_width)
+
+    def set_outline_color(self, color_values: list):
+        color = color_values_to_gdk(color_values)
+        self.outline_color_chooser_button.button.set_rgba(color)
+
+    def set_alignment(self, alignment: str):
+        self.alignment_buttons.set_alignment(alignment)
+
+    def on_change_color(self, _):
+        color = self.color_chooser_button.button.get_rgba()
+        color = gdk_color_to_values(color)
+
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_font_color(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_color=color)
+
+        self.color_chooser_button.revert_button.set_visible(True)
+
+    def on_change_outline_width(self, _):
+        width = int(self.outline_width.button.get_value())
+
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_outline_width(identifier=self.active_identifier, state=self.state, label_position=self.key_name, outline_width=width)
+
+        self.outline_width.revert_button.set_visible(True)
+
+    def on_change_outline_color(self, _):
+        color = self.outline_color_chooser_button.button.get_rgba()
+        color = gdk_color_to_values(color)
+
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_outline_color(identifier=self.active_identifier, state=self.state, label_position=self.key_name, outline_color=color)
+
+        self.outline_color_chooser_button.revert_button.set_visible(True)
+
+    def on_change_font(self, button):
+        # font = self.font_chooser_button.button.get_font()
+        # name, size = self.parse_font_description(font)
+
+        font_desc = self.font_chooser_button.button.get_font_desc()
+        name, size, weight, style = get_values_from_pango_font_description(font_desc)
+
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_font_family(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_family=name, update=False)
+        active_page.set_label_font_style(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_style=style, update=False)
+        active_page.set_label_font_size(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_size=size, update=False)
+        active_page.set_label_font_weight(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_weight=weight, update=True)
+
+        self.font_chooser_button.revert_button.set_visible(True)
+
+    def on_reset_font(self, button):
+        #FIXME: gets called multiple times
+        active_page = gl.app.main_win.get_active_page()
+        #TODO
+        active_page.set_label_font_family(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_family=None, update=False)
+        active_page.set_label_font_size(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_size=None, update=True)
+
+        button.set_visible(False)
+
+    def on_reset_text(self, button):
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_text(identifier=self.active_identifier, state=self.state, label_position=self.key_name, text=None)
+
+        self.update_values()
+
+        button.set_visible(False)
+
+    def on_reset_color(self, button):
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_font_color(identifier=self.active_identifier, state=self.state, label_position=self.key_name, font_color=None)
+
+        button.set_visible(False)
+
+    def on_reset_outline_width(self, button):
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_outline_width(identifier=self.active_identifier, state=self.state, label_position=self.key_name, outline_width=None)
+
+        button.set_visible(False)
+
+    def on_reset_outline_color(self, button):
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_outline_color(identifier=self.active_identifier, state=self.state, label_position=self.key_name, outline_color=None)
+
+        button.set_visible(False)
+
+    def on_change_alignment(self, button):
+        if not button.get_active():
+            return  # Only respond to the button being activated
+
+        alignment = self.alignment_buttons.get_alignment()
+
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_alignment(identifier=self.active_identifier, state=self.state, label_position=self.key_name, alignment=alignment)
+
+        self.alignment_buttons.revert_button.set_visible(True)
+
+    def on_reset_alignment(self, button):
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_alignment(identifier=self.active_identifier, state=self.state, label_position=self.key_name, alignment=None)
+
+        button.set_visible(False)
+
+    def on_change_text(self, entry):
+        text = entry.get_text()
+
+        active_page = gl.app.main_win.get_active_page()
+        active_page.set_label_text(identifier=self.active_identifier, state=self.state, label_position=self.key_name, text=text)
+
+        self.text_entry.revert_button.set_visible(True)
+
+        hide_details = text.strip() == ""
+        self.font_chooser_box.set_visible(not hide_details)
+        self.outline_box.set_visible(not hide_details)
+        self.alignment_box.set_visible(not hide_details)
